OnlineVoting-Django/requirements.txt
altgraph==0.17.4
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==2.3.3
attrs==23.2.0
auto-py-to-exe==2.42.0
bcrypt==4.0.1
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.3
coloredlogs==15.0.1
cryptography==42.0.5
cssselect2==0.7.0
decouple==0.0.7
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.1
Django==5.0.4
django-allauth==65.1.0
django-bootstrap-modal-forms==2.2.0
django-crispy-forms==1.8.1
django-environ==0.9.0
django-filter==2.2.0
django-import-export==3.3.8
django-login-required-middleware==0.4
django-renderpdf==4.0.0
django-widget-tweaks==1.4.5
djangorestframework==3.14.0
Eel==0.16.0
et-xmlfile==1.1.0
filelock==3.0.12
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.38.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
html5lib==1.1
humanfriendly==10.0
idna==3.4
imageio==2.34.1
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy_loader==0.4
lazy-object-proxy==1.4.3
llvmlite==0.42.0
lxml==5.1.0
MarkupPy==1.14
MarkupSafe==2.1.5
mccabe==0.6.1
mod-wsgi==5.0.0
more-itertools==4.1.0
mpmath==1.3.0
mssql-django==1.2
mysqlclient==2.2.4
networkx==3.3
numba==0.59.1
numpy==1.24.2
odfpy==1.4.1
onnxruntime==1.17.3
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
pandas==1.5.3
paramiko==3.2.0
pefile==2023.2.7
Pillow==9.4.0
pip==24.0
platformdirs==2.6.2
pooch==1.8.1
protobuf==5.26.1
pycparser==2.21
pydyf==0.5.0
pyHanko==0.22.0
pyhanko-certvalidator==0.26.3
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
PyJWT==2.9.0
pylint==2.4.4
PyMatting==1.1.12
PyNaCl==1.5.0
pyodbc==4.0.35
pyparsing==3.1.1
pypdf==4.1.0
pyphen==0.13.2
pypng==0.20220715.0
pypyodbc==1.3.6
pyreadline3==3.4.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2019.3
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.35.0
rembg==2.0.56
reportlab==4.0.9
requests==2.31.0
rpds-py==0.18.0
scikit-image==0.23.2
scipy==1.13.0
setuptools==65.5.0
six==1.13.0
sqlparse==0.5.0
svglib==1.5.1
sympy==1.12
tablib==3.5.0
tifffile==2024.4.24
tinycss2==1.2.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2022.7
tzlocal==5.2
Unipath==1.1
uritools==4.0.2
urllib3==2.0.3
virtualenv==20.4.3
weasyprint==58.0
webencodings==0.5.1
Werkzeug==3.0.1
whichcraft==0.6.1
whitenoise==6.6.0
wrapt==1.11.2
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0
zope.event==5.0
zope.interface==6.2
zopfli==0.2.2
